clock

+++ components/native.cpp
@@ -3,7 +3,7 @@
 
 std::any LoxClock::call(Interpreter &ip, const std::vector<std::any> &args) {
   const auto now = std::chrono::system_clock::now();
-  return static_cast<int>(
+  return static_cast<double>(
       std::chrono::duration_cast<std::chrono::seconds>(now.time_since_epoch())
           .count());
 }
